CMCASETASK-536: RUP frame No has been added to diagnostics output of TuningService.

+++ Simulator/SimSoftware.cpp
@@ -120,7 +120,7 @@ namespace  Sim
 	QStringList Software::monitors() const
 	{
 		QStringList result;
-		result.reserve(m_software.size());
+		result.reserve(static_cast<int>(m_software.size()));
 
 		for (const auto& app : m_software)
 		{
